+++ packages/openapi-ts/src/plugins/@pinia/colada/mutationOptions.ts
@@ -3,14 +3,16 @@ import type ts from 'typescript';
 import type { GeneratedFile } from '../../../generate/file';
 import type { IR } from '../../../ir/types';
 import { tsc } from '../../../tsc';
-import {
-  createOperationComment,
-  isOperationOptionsRequired,
-} from '../../shared/utils/operation';
+import { createOperationComment } from '../../shared/utils/operation';
 import { handleMeta } from './meta';
 import type { PluginState } from './state';
 import type { PiniaColadaPlugin } from './types';
-import { useTypeData, useTypeError, useTypeResponse } from './utils';
+import {
+  getPublicTypeData,
+  useTypeData,
+  useTypeError,
+  useTypeResponse,
+} from './utils';
 
 const mutationOptionsType = 'UseMutationOptions';
 
@@ -49,6 +51,10 @@ export const createMutationOptions = ({
   const typeData = useTypeData({ file, operation, plugin });
   const typeError = useTypeError({ file, operation, plugin });
   const typeResponse = useTypeResponse({ file, operation, plugin });
+  const { isNuxtClient, strippedTypeData } = getPublicTypeData({
+    plugin,
+    typeData,
+  });
 
   const identifierMutationOptions = file.identifier({
     $ref: `#/pinia-colada-mutation-options/${operation.id}`,
@@ -111,9 +117,12 @@ export const createMutationOptions = ({
         async: true,
         multiLine: true,
         parameters: [
-          {
-            name: fnOptions,
-          },
+          isNuxtClient
+            ? {
+                name: fnOptions,
+                type: `Partial<${strippedTypeData}>`,
+              }
+            : { name: fnOptions },
         ],
         statements,
       }),
@@ -129,11 +138,6 @@ export const createMutationOptions = ({
     });
   }
 
-  const isRequiredOptionsForMutation = isOperationOptionsRequired({
-    context: plugin.context,
-    operation,
-  });
-
   const statement = tsc.constVariable({
     comment: plugin.config.comments
       ? createOperationComment({ operation })
@@ -142,13 +146,14 @@ export const createMutationOptions = ({
     expression: tsc.arrowFunction({
       parameters: [
         {
-          isRequired: isRequiredOptionsForMutation,
+          isRequired: false,
           name: 'options',
-          type: typeData,
+          type: `Partial<${strippedTypeData}>`,
         },
       ],
-      // TODO: better types syntax
-      returnType: `${mutationOptionsType}<${typeResponse}, ${typeData}, ${typeError.name}>`,
+      returnType: isNuxtClient
+        ? `${mutationOptionsType}<${typeResponse}, ${strippedTypeData}, ${typeError.name}>`
+        : `${mutationOptionsType}<${typeResponse}, ${typeData}, ${typeError.name}>`,
       statements: [
         tsc.returnStatement({
           expression: tsc.objectExpression({